easy first refactors creating the structure
